requirements.txt
absl-py==2.1.0
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.11.0
aiosignal==1.3.1
albucore==0.0.13
albumentations==1.4.14
altair==5.4.1
annotated-types==0.7.0
ansitable==0.11.3
anyio==4.6.2.post1
async-timeout==5.0.1
attrs==24.2.0
autodistill==0.1.28
autodistill-grounding-dino==0.1.4
bidict==0.23.1
blinker==1.9.0
cachetools==5.5.0
cffi==1.17.1
cfgv==3.4.0
clip @ git+https://github.com/ultralytics/CLIP.git@61a901215e3e7c706a2f32b5f9969f9dec2beaf9
colorama==0.4.6
colored==2.2.4
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dirtyjson==1.0.8
eval_type_backport==0.2.0
evdev==1.7.1
exceptiongroup==1.2.2
executing==2.1.0
fairscale==0.4.13
ffmpy==0.4.0
filetype==1.2.0
Flask-Cors==5.0.0
Flask-SocketIO==5.4.1
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.5.0
ftfy==6.3.1
gdown==5.2.0
greenlet==3.1.1
h11==0.14.0
-e git+ssh://git@github.com/Techolution/hi_robotics.git@092d138317f193bd2dbcc907ce65ac735f2fecb0#egg=hi_robotics
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.3
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
ipykernel==6.29.5
itsdangerous==2.2.0
jax==0.4.35
jax-cuda12-pjrt==0.4.34
jax-cuda12-plugin==0.4.34
jaxlib==0.4.35
joblib==1.4.2
jsonschema==4.23.0
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.7
lazy_loader==0.4
lightning-utilities==0.11.6
llama-index-core==0.11.23
llvmlite==0.43.0
markdown-it-py==3.0.0
marshmallow==3.23.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.18
moondream==0.0.5
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.8.3
nest-asyncio==1.6.0
nltk==3.9.1
nodeenv==1.9.1
numba==0.60.0
numpy==2.2.2
nvidia-cuda-nvcc-cu12==12.6.77
nvidia-ml-py==12.535.161
nvidia-nvjitlink-cu12==12.4.127
nvitop==1.3.2
onnx==1.17.0
onnxruntime==1.20.1
openai-whisper==20231117
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
packaging==24.2
paho-mqtt==2.1.0
pandas==2.2.3
parso==0.8.4
pillow==11.1.0
pre_commit==4.0.1
progress==1.6
propcache==0.2.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==17.0.0
pycocotools==2.0.8
pycparser==2.22
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
pynput==1.7.7
pyparsing==3.2.0
pyrealsense2==2.55.1.6486
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.10.1
python-socketio==5.11.4
python-xlib==0.33
pytorch-lightning==2.4.0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rf-groundingdino==0.1.2
rf-segment-anything==1.0
roboflow==1.1.48
ruff==0.6.9
safetensors==0.4.4
scikit-image==0.24.0
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.2.0
shellingham==1.5.4
simple-websocket==1.1.0
sniffio==1.3.1
sounddevice==0.5.1
spatialmath-python==1.1.13
SpeechRecognition==3.10.4
SQLAlchemy==2.0.36
stack-data==0.6.3
starlette==0.41.3
streamlit==1.38.0
supervision==0.24.0
sympy==1.13.1
tenacity==8.5.0
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2024.8.24
tiktoken==0.8.0
timm==1.0.9
tokenizers==0.20.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.0
torchmetrics==1.4.1
tornado==6.4.1
traitlets==5.14.3
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.14
ultralytics-thop==2.0.11
uvicorn==0.32.1
watchdog==4.0.2
Wave==0.0.2
wcwidth==0.2.13
websocket-client==1.8.0
websockets==14.2
wrapt==1.16.0
wsproto==1.2.0
yapf==0.40.2
yarl==1.17.1
